requirements.txt
panel>=1.5.0
holoviews>=1.19.0
hvplot>=0.11.0
geoviews>=1.13.0
datashader>=0.16.0
bokeh>=3.5.0
pandas>=2.2.0
numpy>=1.26.0
param>=2.1.0
cartopy>=0.23.0
pyproj>=3.6.0
shapely>=2.0.0
requests>=2.32.0
python-dotenv>=1.0.1
watchfiles
pydeck>=0.9.1
websocket-client>=1.7.0
